[markdown] Update SmartyPants Lib to 1.8.1.
